// src/app/models/cultivo.interface.ts
export interface CultivoI {
    id?: string;
    tipo: string;
    cantidad: number;
    ubicacion: string;
    fechaI: string;
    nombreE: string;
    humedadMax: number;
    humedadMin: number;
    oxigenoMax: number;
    oxigenoMin: number;
    temperaturaMax: number;
    temperaturaMin: number;
    phMax: number;
    phMin: number;
    nombreP: string;
    humedadMaxP: number;
    humedadMinP: number;
    oxigenoMaxP: number;
    oxigenoMinP: number;
    temperaturaMaxP: number;
    temperaturaMinP: number;
    phMaxP: number;
    phMinP: number;
}